Refactor: Move test endpoints to a dedicated router
This commit resolves a `NameError` that occurred during application startup. The error was caused by having API endpoints defined directly in the main application file, which can lead to circular dependencies or initialization order issues.

The solution involves:
1. Creating a new router file `kpi_dashboard/backend/app/routers/test.py`.
2. Moving the test-related endpoints (`/api/test-celery-task`, `/api/test-endpoint`) from `kpi_dashboard/backend/app/main.py` to the new router file.
3. Updating `main.py` to import and include the new test router.

This change follows FastAPI best practices by separating concerns and ensures that the main application object is fully initialized before any routes are registered, thus preventing the `NameError`. The application now starts up correctly, as verified by running the test suite.
